Split sync delta observability
