+++ dev-docs/module-reference.md
@@ -1403,7 +1403,11 @@ Not separately documented because their exports are best read from source — ke
 - `sync-delta-surfaces.js` — DELTA_ARRAYS/MAPS/SCALARS importedData surface lists for per-row sync.
 - `sync-delta-surface-config.js` — per-surface itemId/keyId overrides for array/map delta planners and pull overlays.
 - `sync-delta-id.js` — shared per-row identity helpers, stable hash, allowlist guard, and proto-pollution key rejection.
-- `sync-delta-observability.js` — per-push delta telemetry, pull-side row-count snapshots, and Phase 2 cutover readiness checks.
+- `sync-delta-observability.js` — delta observability facade; compatibility re-exports for telemetry, pull snapshots, readiness, and injected Evolu query access.
+- `sync-delta-observability-context.js` — shared Evolu/itemRow query dependency access for delta observability helpers.
+- `sync-delta-pull-snapshot.js` — in-memory pull-side row-count snapshots for Sync Diagnose delta replication checks.
+- `sync-delta-telemetry.js` — per-push delta telemetry storage, aggregation, reset, and latest pull-snapshot reporting.
+- `sync-delta-readiness.js` — Phase 2 cutover readiness checks across array/map/scalar delta surfaces.
 - `sync-apply.js` — inbound sync apply helpers for AI provider settings and display prefs, with compatibility re-exports for chat apply helpers.
 - `sync-chat-apply.js` — inbound chat thread/message application, chat delete tombstones, and local freshness locks that protect just-edited local chat from stale remote rows.
 - `sync-tombstones.js` — remote profile delete propagation helpers, inbound tombstone application, batched-delete quarantine, and apply/reject pending tombstone actions for Settings.
