+++ service-worker.js
@@ -165,6 +165,10 @@ const APP_SHELL = [
   '/js/sync-delta-surface-config.js',
   '/js/sync-delta-id.js',
   '/js/sync-delta-observability.js',
+  '/js/sync-delta-observability-context.js',
+  '/js/sync-delta-pull-snapshot.js',
+  '/js/sync-delta-telemetry.js',
+  '/js/sync-delta-readiness.js',
   '/js/sync-tombstones.js',
   '/js/sync-messenger.js',
   '/js/sync-environment.js',
